unity build

+++ WinApiLecture/Client/CPathMgr.cpp
@@ -13,7 +13,7 @@ void CPathMgr::init()
 {
 	GetCurrentDirectory(255, m_szContentPath);
 
-	int iLen = wcslen(m_szContentPath);
+	int iLen = (int)wcslen(m_szContentPath);
 
 	for (int i = iLen - 1; i >= 0; --i)
 	{
